Fix column ordering bug in HashJoin filter
In the HashJoin filter if the build columns appeared intermixed with
the probe columns, there was a bug in how the ROW structure for the
filter evaluation was built. It presumed that all the probe columns
come before the build columns which resulted in the channels being
wired differently for picking values during filter evaluation.This
caused errors and crashes if the types of the columns in the same
channel position of the input tables had incompatible types.

This bug showed up in TPC-H query 19 where the build and probe side
columns were intermixed in the filter it used.