Fix DriverTest end of result set
DriverTest.pause and possibly other tests can be flaky so that the end
state of the task may not be kFinished and the final Driver count may
not be 0. The flakiness happens because destruction of TaskCursor
calls requestTerminate on the Task after the end of the result set. At
this time, there can b threds inside the Task. If so, these take an
error path to exit and may leave the state to be kCancelled or
kFailed.  The fix is to request terminate on the Task tat TaskCursor
destruction only if there is unread output. If all output is consumed,
the Task will go off thread and transit to kFinished in a short time.

To check that the end state is correct, we take a stateChangeFuture()
and wait for a non-running state, which is the official API for
listening to state changes.