Add Reader for Variadic Type (#701)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/701

This diff adds an implementation of VectorReader<Variadic<T>>, which returns
a VariadicView<T> for accessing the values of individual arguments.

VariadicView<T> exposes an interface similar to std::vector and heavily based
on ArrayView<T>.  VariadicView<T> maintains  pointers to a collection of
readers and allows access to the value from each of these readers at the
current offset, either directly or by iterating over the readers in turn.

I split out DedodedArgs into its own header file because VectorReader<Variadic<T>>
takes an instance of DedodedArgs in its constructor.  Adding the Expr.h header to
VectorUdfTypeSystem.h caused issues in presto_cpp because Expr.h includes the
Expressions.h header which calls the isRow() function on a Type in one of its functions.
 presto_cpp defines a macro isRow() which conflicts with this function name.  Splitting out
DedodedArgs into its own header file and including that in Expr.h works around this issue.

Another option would be to pass in a vector<DecodedVector*> instead of DecodedArgs
(which would be nicer in other ways too), but comes at the cost of having to build that vector.

I will integrate the Reader into UDFs in a following diff.

Reviewed By: laithsakka

Differential Revision: D32840896

fbshipit-source-id: b63a1af9a6d06e28ea2cf747ef103af1fc76952c