Fix bug in FunctionSignature's parseTypeSignature
Summary:
There's a bug in FunctionSignature's parseTypeSignature function where it will not respect complex types when looking for the end of nested types.

For example:
row(map(K1, V1), map(K2, V2))

when parsing the inner types of the struct, it will parse map(K1, V1) just fine (recognizing the comma is part of the map's type signature), but when looking for the end of map(K2, V2) it will treat the inner comma as a separator between types in row's signature, and try to parse map(K2 as a type.  This of course leads to an error in parsing.

The fix is to call findNextComma in the for loop instead of signature.find(',' since findNextComma respects commas that are part of inner types.

Reviewed By: kgpai, pedroerp

Differential Revision: D33110432

fbshipit-source-id: 141440945916ddb0681c94e30167d73a629c4525