Abandon low value dictionary under memory pressure (#703)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/703

Currently in velox writer, we determine whether or not to use dictionary encoding right before the flush, but the encoding conversion introduces significant memory overhead and complicates memory management.

Instead, we now try to abandon low value dictionaries before we try to flush. This increases the granularity of writer memory increment, which both reduces the overhead we need to reserve for memory management and also allows us to delay premature flushes as much as possible without changing the current flow.

NOTE: `abandonDictionaries` can be called even after the first stripe and would do nothing in that case. `abandonDictionaries(false)` doesn't force abandon the dictionary and is natively used to determine whether to use dictionary encoding. The method exits early when we already determined not to use dictionary encoding for the column reader instance.

Reviewed By: zzhao0

Differential Revision: D32874984

fbshipit-source-id: 761106d1f7149e413e67b61d8b752a1740b7ac78