Refactor reading ctor to construct ek internally (#666)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/666

Now that we added flatmap context as a member of ColumnReader, flatmap info like sequence should be stored only in flatmap context.

This called for removing EncodingKey member and using it only as a utility class for getting stream identifiers and we only construct the classes as needed (usually in builder and ctor body).

Reviewed By: zzhao0

Differential Revision: D32574843

fbshipit-source-id: c18d06aa3dcc01fa998a48ce09abb84330610b3d