Support Equality and Hashing for FunctionSignatures (#747)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/747

**High Level Goal:**
I want to change ScalarFunctions and AdaptedVectorFunctions to
do function call lookups based on FunctionSignatures.

This is primarily because the current solution of doing exact match
lookups against the argument types is overly restrictive and makes
adding features like VariadicArgs difficult.  SignatureBinder is much
 more flexible and has the flexibility to allow for more features in the
future.

With VectorFunctions already using FunctionSignatures, and Aggregate
Functions in the process of migrating, this also represents a step towards
 consistency in behavior.

**This change:**
This change adds support for hashing and equality operators for
FunctionSignatures, so we can used them as keys in unordered
collections.

Vector and Aggregate functions register all schemas at the same time.
Scalar and AdaptedVector functions allow users to register different
schemas for the same function name in different calls.

In order to allow for this, while also efficiently ignoring duplicate signatures,
I plan to use an unordered_map from FunctionSignature to Factory in
the new registration system, which requires that FunctionSignatures
support hash and equality.

Reviewed By: laithsakka

Differential Revision: D33077656

fbshipit-source-id: 670f9695e531f429e11926b14dc64f9d6fb0c9b1