Add streaming aggregation (#771)
Summary:
Add streaming aggregation operator that can be used when input is already
pre-grouped on the grouping keys. This operator is using less memory than hash
aggregation because it doesn't need to accumulate all groups in memory. The
input must be grouped on the grouping keys, but doesn't need to be sorted on
these keys.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/771

Reviewed By: oerling

Differential Revision: D33197699

Pulled By: mbasmanova

fbshipit-source-id: ed274fe394d8b16d17619e2f126beb3b0d59a2dc