Introduce FunctionRegistry for Signature based registration of Scalar/AdaptedVector Functions (#749)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/749

**High Level Goal:**
I want to change ScalarFunctions and AdaptedVectorFunctions to do
function call lookups based on FunctionSignatures.

This is primarily because the current solution of doing exact match
lookups against the argument types is overly restrictive and makes
adding features like VariadicArgs difficult.  SignatureBinder is much
more flexible and has the flexibility to allow for more features in the
future.

With VectorFunctions already using FunctionSignatures, and Aggregate
Functions in the process of migrating, this also represents a step towards
consistency in behavior.

**This change:**
This change introduces the FunctionRegistry which will handle registration
and lookups for Scalar and AdaptedVector Functions. &Note, it's not used yet,
that will come in future diffs). At it's heart, it's just a wrapper around a map
from function name to a map of signature to factory.

The reason for the inner map is to allow quick deduping when an identical
function signature is registered (which unfortunately seems to happen
frequently).

It exposes the same sort of functions ScalarFunctionRegistry exposed,
except as class member functions rather than raw top level functions.
Notably registerFunction and createFunction.  In addition, to support
cases that need to expose available functions, it supports fetching all
available function names, and all available signatures for a given name.

The major difference with ScalarFunctionRegistry is that it uses
SignatureBinder to find the corresponding signature for a given function
 call, rather than checking strict argument equality.

Reviewed By: laithsakka

Differential Revision: D33078596

fbshipit-source-id: a12d5696c640b66e3b551d86ce20062baeecd03d