Efficient crc computation by skipping double memory allocation in PrestoSerializer (#608)
Summary:
Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/608

Pull Request resolved: https://github.com/facebookincubator/velox/pull/781

In PrestoSerializer flush method, we compute CRC with incoming data. This data is copied to a string stream, and then CRC is computed. This is not efficient, as we already walk through the bytes during flush. In this diff, we create a "OutputStream" class, which works as a proxy on std::ostream. We also provide a OutputStreamListener class that is used as a callback for write method. PrestoOutputStreamListener class inherits default outputstream listener and extends to support crc computation when a write is called on OutputStream.

Reviewed By: mbasmanova

Differential Revision: D33231829

fbshipit-source-id: 3e54d622e6e4c1809d3f143716be63689b8ce73a