Add SsdFile - A Shard of a SSD cache for AsyncDataCache
Entries from AsyncDataCache can be written to and read from an
SsdFile. When writing, the file allocates space for the cache entries,
possibly growing the file or evicting older entries.  When reading,
the caller obtains SsdPins for the entries to be read. The entries are
guaranteed to stay live for the duration of their pin. The entries can
then be read with a coalesced IO.

The SsdFile is composed of regions. Each region has entries and
entries do not span regions. The region is the unit of eviction and
pinning and access tracking. When the cache is read from, region read
counts are updated and older counts are decayed. When needing to write
to a full SsdFile, we take the lowest score unpinned region and remove
the entries in this region from the key to entry map and rewrite the
region.

File contents are not kept across restarts. A checkpointing mechanism
that periodically makes a snapshot of cache state can be added in the
future.

Failure to read throws an error. Failure to write logs the error and
leaves a state where the entry is not marked written. TBD: Make a
generic means for issuing an alert.