Minor fix for AbstractJoinNode error message (#810)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/810

This is a minor fix for `AbstractJoinNode` error message, changing usage of `probe` and `build` to `left` and `right`, as we have both hash join and merge join as subclass of `AbstractJoinNode`. `probe` and `build` is just terminology working for hash join, but not for merge join. This helps to give more clearer error message for merge join.

Differential Revision: D33304451

fbshipit-source-id: a909d5aa64cb8bde85f78d5aff2575c63693e462