Remove static driver executor (#820)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/820

Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/612

Velox used to have default cpu executor. We've decided to remove that and instead have it supplied by user through QueryCtx. Executor is not required if velox is only used for expression eval, so we still have default nullptr in the constructor. The check against null happens at that time when the executor is used by Driver.

Reviewed By: mbasmanova

Differential Revision: D33299548

fbshipit-source-id: 664effbab0d64a9683cf06bf060505add4d0d167