Fix rtrim for strings with invalid UDF-8 characters
Summary:
Modify rtrim function to traverse input string backwards to correctly handle
invalid encodings and improve performance.
Before this change, rtrim used to traverse the string from the start looking for
the last non-whitespace character. The complexity of that algorithm is O ( N )
 while the complexity of the new algorithm is O(number of whitespace
characters at the end of the string).
Also, the previous version used to stop processing if there was an incorrectly
encoded character in the middle of the string, e.g. it would not trim white
spaces at the end of such strings. This behavior doesn't match Presto.
While working on this change, I noticed that the set of supported whitespace
characters is less than the set supported in Presto. A follow-up diff will
expand the set to match Presto.

Reviewed By: mbasmanova

Differential Revision: D33037624

fbshipit-source-id: 9d3f4c1483c0664377a7033bf415d8988b9ac7e0