Support getting a FunctionSignature from IScalarFunction (#750)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/750

**High Level Goal:**

I want to change ScalarFunctions and AdaptedVectorFunctions to do function call
 lookups based on FunctionSignatures.

This is primarily because the current solution of doing exact match lookups against
the argument types is overly restrictive and makes adding features like VariadicArgs
difficult.  SignatureBinder is much more flexible and has the flexibility to allow for more
features in the future.

With VectorFunctions already using FunctionSignatures, and AggregateFunctions in
the process of migrating, this also represents a step towards consistency in behavior.

**This change:**
The goal of this change is just to add a signature() function to IScalarFunction that will
 return a FunctionSignature object for the given Function.

FunctionSignature is under the velox_expression library and IScalarFunction in ScalarFunction in
velox_core.  Since velox_expression depends on velox_core this introduces a circular dependency.

To resolve this I've moved FunctionSignature and a couple related files (SignatureBinder, FunctionRegistry) into a separate velox_expression_functions library which does not depend on velox_core.  Since velox_core and velox_expression can both depend on velox_expression_functions there are no further changes needed to other files.

Reviewed By: HuamengJiang

Differential Revision: D33078846

fbshipit-source-id: c80b8ffac2fe721e4a4b746dd0bc439b610feaf4