Simplify FusedLoad and CachedBufferedInput
The IO pattern for Velox file readers is to register the pieces of
file to be accessed, then to access some or all of these
pieces. Pieces that are in close proximity get loaded with a single
IO. The definition of close depends on the media.

This removes the load_ member from AsyncDataCacheEntry. The cache entry no longer references the FusedLoad that will be invoked to fill it. Instead, the CacheInputStream  that will load a piece of file coalesced with other pieces references the FusedLoad.

The API contract of CachePin is simplified: If findOrCreate returns an
exclusive pin, the caller is for filling the entry and calling
setExclusiveToShared when done. This makes the data visible to other
threads. There is no more a concept of dataValid or ensureLoaded for
shared mode pins that have uninitialized data.

We add tests for checking a consistent after state for errors in
loading data into CachePins. AsyncDataCacheTest adds error injection
in FusedLoads ad single pin loading. The code is also more
representative of the the intended usage pattern.

CachedBufferedInput is adapted to use the changed AsyncDataCache
API. CacheInputStream looks for its FusedLoad instead of the Fusedload
being recorded in a cache entry. Cache entries for FusedLoads are made
right before loading the FusedLoad. CoalesceIO.h is used
systematically to detect and coalesce coalesced loads.

ScanTracker exports TrackingData so that columns can be more
accurately sorted in different load frequency buckets.