Fix Expr bug on peeling encoding for single-arg deterministic func
Summary:
Motivation:
A crash was hit when running expression `not(contains(...))` and the root cause is nullptr reference in casting the input to flat vector. This indicates that a bug on argument encoding peeling of `Expr`.

Fix:
The reproduce scenario is turned into a test case of ExprTest where the input is dictionary vector but with `DictionaryVector::isConstant(rows)` equals to true. We handled this by creating a ConstantVector of out it so that the further execution can go into `applySingleConstArgVectorFunction`.

Reviewed By: mbasmanova

Differential Revision: D33523786

fbshipit-source-id: 8ed95f8a44d334395980bb0166f175db9c49fe91