Use Signature Binding to Register Scalar/AdaptedVector Functions (#752)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/752

**High Level Goal:**

I want to change ScalarFunctions and AdaptedVectorFunctions to do function call lookups
based on FunctionSignatures.

This is primarily because the current solution of doing exact match lookups against the
argument types is overly restrictive and makes adding features like VariadicArgs difficult.
SignatureBinder is much more flexible and has the flexibility to allow for more features in
 the future.

With VectorFunctions already using FunctionSignatures, and AggregateFunctions in the
process of migrating, this also represents a step towards consistency in behavior.

**This change:**
This diff switches over from the old equality based Registry to the new signature-based
FunctionRegistry.

The bulk of the change is modifying ScalarFunctions() and AdaptedVectorFunctions() to
return the new objects, and modifying call sites to call the member functions instead of
the top level raw functions.

In addition, I deleted the now irrelevant code from ScalarFunctionRegistry.h and
VectorFunctionBindings.h  (I had to pull GetSingletonUdfMetadata up out of the
FunctionRegistry class as I realized we have dependencies outside Velox using
this function once I deleted it from ScalarFunctionRegistry.h)

We can also probably delete some more code from ScalarFunction.h (e.g. ArgumentsCtx
and FunctionKey), but I'll do that in a later diff.

Reviewed By: tgalkovskyi, laithsakka

Differential Revision: D33079339

fbshipit-source-id: be57ac839892c20d1fae2648dc0234c43f09ee51