Support Variadic Types in Simple Functions (#715)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/715

This diff finishes off the stack to support allowing the list argument in a Simple
Function to be Variadic (a call can have any number of arguments of that type).

The biggest change here is to ArgumentsCtx which is used as a part of FunctionKey
to do lookups of functions based on the name and argument types passed to that
function.  Prior to this diff, argument Types and numbers were fixed, so lookups could
be done by hashing the function name and the types of the arguments in an undordered_dict.

With Variadic arguments, I changed FunctionKey to only hash the function name.
I also changed ArgumentsCtx to treat other instances that are compatible as equal.
For example, ArgumentsCtx({Varchar, Varchar}), ArgumentsCtx({Varchar, Varchar,
Variadic(Varchar)}), ArgumentsCtx({Varchar, Variadic(Varchar)}), and ArgumentsCtx({
Variadic(Varchar)}) are all nowconsidered equal. This is necessary because a function
call will always have well defined arguments, but the function definition may have a
Variadic arg.

In addition, I added some logic to UDFHolder to validate that only the last argument
may be Variadic, and some logic to ScalarFunctionMetadata to expose whether or
not it has a Variadic argument (mainly for the construction of FunctionKeys).

Lastly because the constructor for VectorReader<Variadic<T>> needs to take in
additional arguments (because it needs to construct the VectorReaders for all the
underlying args), I modified unpack in VectorFunctionAdapter to handle this special case.

Reviewed By: laithsakka

Differential Revision: D32934217

fbshipit-source-id: d2e9f2c5c061ca9ed33ae1a89c240b34b2198007