Fix incorrect assert in SelectiveColumnReader and add seeds to BatchMaker::createBatch() (#889)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/889

BatchMaker::createBatch() internally creates a std::mt19937 object to generate
random numbers, however, each time it is called it uses the same default seed,
this causes it to generate multiple same batches, which makes it much less
useful to test multiple batches. This diff allows the caller to pass a seed
to this function to generate differet batches.

When using this fixed version of createBatch() in E2EFilterTest, I found that
one of the assert in SelectivecColumnReader was not correct, and was causing
crash. So fixed that as well.

Reviewed By: zzhao0

Differential Revision: D33590384

fbshipit-source-id: 00a65e5b7f2a6e3fa23c88d85a131620d2d63eb8