Add a callback for increasing a memory cap
A MemoryUsageTracker can try to raise its limit instead of throwing an
exception. Child and parent trackers can in principle both have
caps. First we check against the cap of parents, then of children. If
the parent raise succeeds and the child raise fails, the parent raise
stays in effect. The grow callback is called with the tracker set as
if the allocation that exceeded the limit succeeded. The callback may
raise the limit to accommodate the allocation or return false to
indicate that the allocation should be reverted and an exception
thrown.

The practical use case has a Task level tracker with a cap for each
Task.  The grow callback of this will look at other Tasks and see if
growing is possible or if Tasks should spill or be killed.