Remove ScalarFunctions (#888)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/888

While working with SimpleFunctions I found the code difficult to work with,
primarily because SimpleFunctions are either referred to as ScalarFunctions
or AdaptedVectorFunctions in the code. I also found myself doing a lot of
duplicate work for the ScalarFunctions and AdaptedVectorFunctions code paths.

In the code itself, it looks like ScalarFunctions are only ever used to access
the metadata for SimpleFunctions, while AdaptedVectorFunctions are used mainly
 for execution.

To try to simplify this, I've made the following changes:
1) I removed the ScalarFunctionRegistry and renamed AdaptedVectorFunctionRegistry to
SimpleFunctionRegistry, it is now the only place simple functions are registered.

2) FunctionRegistry is now templated by both the "UDF" type and the metadata type, it also
now returns a FunctionEntry which can either return an object of the metadata type or the
"UDF" type (the "UDF" is currently SimpleFunctionAdapterFactory for Simple functions),
this means we no longer need to instantiate UDFs to access their metadata from the registry.

3) A bunch of renaming, basically I replaced Scalar and AdaptedVector with Simple.

For (2) I also considered exposing the metadata from the SimpleFunctionAdapterFactory
interface and keeping the FunctionRegistry return type the same.  My thinking for not
doing this was that a) it felt weird to make a factory object stateful like that and b) that
meant we'd have to instantiating a SimpleFunctionAdapterFactoryImpl every time we
wanted to access the metadata.

Reviewed By: laithsakka

Differential Revision: D33511875

fbshipit-source-id: 8fa3267eb2fade6c9c210c14bea42346280a4fd4