Fix Task initialization order
Declare queryCtx_ ahead of pool_. Move pool_ and queryCtx_ up in
declaration order so that pool_ is before items that may be allocated
from this, meaning that pool_ will be destructed after the items.
Move pool_ initialization in constructor to match declaration
order. Initializers are generally expected to reflect the order of
declaration.

Previous Task construction crashes at least
with gcc 10.3.1.