Adds an SSD layer  for AsyncDataCache
If AsyncDataCache is given an optional SsdCache, moves qualifying
entries into the SsdCache as a by-product of scanning entries for
eviction.

SsdCache has the same key and quantization as the RAM cache. Writing
to SSD is done on a best effort basis so as not to block reading from
storage while waiting to save on SSD.

SsdCache is a container for multiple shards, represented by
SsdFile. Each shard has its map from SsdKey, a file id and offset, to
SsdRun, an offset and size in the file of the shard. Extra metadata
can be added to the map to represent for example recoding of data. An
SsdPin is the result of looking up an entry. The entry will stay
readable from the file of the shard for the lifetime of the pin.

The shard for an entry is selected on a hash of the storage file of
the entry. In this way correlated entries stay together and load
coalescing applies to SSD as well as to storage. We expect large
numbers of distinct files and expect no significant skew.

Adds a post-load hook to AsyncDataCache for checking integrity of read data.