Add SsdCache to CachedBufferedInput
Adds wiring to CacheInputStream and CachedBufferedInput for reading
from SsdCache where possible.  Adds accounting of file and file group
level access to ScanTracker/FileGroupStats. FileGroupStats is is a
placeholder for logic for deciding which data to stage on SSD.

CachedBufferedInput divides enqueued regions between RAM, SSD and
storage. Such a division is made for different access frequency
buckets. Columns that are accessed at the same frequency are assumed
to be correlated. This is a reasonable assumption in uses that do a
conjunction of column filters. Loads are coalesced within these
buckets. Loads of the most densely accessed bucket are scheduled on an
executor for prefetching if the executor is supplied. Other loads are
made on first use, combining all nearby loads in one IO.

When accessing a single addressable unit, e.g. values or nulls of a
column, we look up the entry n the RAM cache, then in the SSD cache
and finally load from storage. An entry can be cached with a different
length than requested if we have loadss using a different load
quantum. If the entry that is found in a cache is shorter than the
requested size, the entry is ignored and will be replaced by a fresh
longer entry. If loading an entry rom SSD fails, we log the error and
proceed to load the data from storage, We remove said entry from SSD
so as not to repeatedly hit the same error.