Fix cardinality of Boolean in VectorHasher
Fixes a bug with array/normalized key aggregation with boolean keys
When the data is all false for eth creation of the hash table, the
column has a cardinality of 1 by max-min. When a true value comes
along, this is mappd to a constant without looking at the min-max and
indices go out of range. So a boolean key will have a cardinality of 3
for either range or value ids. A boolean key is always expected to
have a range mapping. cardinality() was correct but enableValueRange()
was not consistent with it.